% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchdamic_new_methods.R
\name{ancombc}
\alias{ancombc}
\title{ANCOMBC}
\usage{
ancombc(ps, formula, group)
}
\arguments{
\item{ps}{A phyloseq object}

\item{formula}{A string.}

\item{group}{A string.}
}
\value{
ANCOMBC result.
}
\description{
\code{ancombc} performs ancombc
}
