% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{counts_to_relative_abundance}
\alias{counts_to_relative_abundance}
\title{Convert matrix of counts to relative abundance}
\usage{
counts_to_relative_abundance(x, total_reads = NULL, total_sum = 100)
}
\arguments{
\item{x}{A numeric matrix of counts. Features in rows and samples in columns.}

\item{total_reads}{Optional. A character vector with the total raw reads/sums
of the matrix per column. If not provided (i.e. NULL), the total sum per
column will be used.}

\item{total_sum}{The scaling factor. Default 100 (percent).}
}
\value{
A matrix of relative abundance.
}
\description{
\code{counts_to_relative_abundance} converts a matrix of counts to
relative abundance
}
