% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_DA_methods.R
\name{DA_wilcox}
\alias{DA_wilcox}
\title{Wilcox test for differential abundance}
\usage{
DA_wilcox(
  object,
  pseudo_count = FALSE,
  norm = "none",
  conditions_col,
  conditions,
  verbose = FALSE
)
}
\arguments{
\item{object}{A phyloseq object.}

\item{norm}{String character. Choose between three normalization methods:
'none', 'CLR', or 'TSS'.}

\item{conditions_col}{String character.
The name of the column in sample data with the conditions.}

\item{verbose}{Print messages or not, Default is FALSE.}
}
\value{
A list with outputs compatible with the benchdamic framework.
}
\description{
\code{DA_wilcox} performs Wilcoxon test on a phyloseq object.
}
