% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_methods_new.R
\name{DA_wilcox_test}
\alias{DA_wilcox_test}
\title{Wilcox test}
\usage{
DA_wilcox_test(object, grp, ref = NULL, ...)
}
\arguments{
\item{object}{A phyloseq or (Tree)SummairizedExperiment object}

\item{grp}{A character string with the column name.}

\item{ref}{A character string with the condition used as reference.}

\item{...}{Other paramteres}
}
\value{
A list compatibble with the benchdamic workflow.
}
\description{
\code{wilcox_test} is an adaptation of the wilcox.test function
(stats package) for use with the benchdamic workflow.
}
