% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_DA_methods.R
\name{DA_ZINQ}
\alias{DA_ZINQ}
\title{ZINQ for benchdamic}
\usage{
DA_ZINQ(
  object,
  pseudo_count = FALSE,
  conditions_col,
  conditions,
  norm = "none",
  pval_method,
  y_CorD,
  verbose = FALSE
)
}
\arguments{
\item{object}{A (Tree)SummarizedExperiment or a phyloseq object.}

\item{pseudo_count}{Whether include or not a pseudo_count. Default = FALSE.}

\item{conditions_col}{The name of the grouping column, which is located in
the colData (SummarizedExperiment) or the sample_data (phyloseq).}

\item{conditions}{A character vector indicating the conditions. Example:
\code{c(condB = 'control', condA = 'treatment')}}

\item{norm}{A character string indicating the normalization method to be used
prior analysis. Default is 'none'.}

\item{pval_method}{A character string indicating the type of pvalue to use.
Options: Cauchy or MinP. Default: Cauchy.}

\item{y_CorD}{Indicate if data are counts 'D' or continuous 'C'.}

\item{verbose}{Whether include messages or not. Default is FALSE.}
}
\value{
A list in the format used in the benhcdamic pipeline.
}
\description{
\code{zinq} adapts the zinq method for integration with the \code{benchdamic}
package.
}
\section{References}{

Ling, W., Zhao, N., Plantinga, A.M. et al. Powerful and robust non-parametric
association testing for microbiome data via a zero-inflated quantile approach
(ZINQ). Microbiome 9, 181 (2021). https://doi.org/10.1186/s40168-021-01129-3
}

