% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{editMiaTaxaNames}
\alias{editMiaTaxaNames}
\title{Edit the taxa names by mia}
\usage{
editMiaTaxaNames(x)
}
\arguments{
\item{x}{A TreeSummarizedExperiment agglomerated by the \code{agglomerateByRank}
function of the mia package.}
}
\value{
A character vector with new names
}
\description{
\code{editMiaTaxaNames} corrects the taxonomic names of the
\code{agglomerateByRank} function of the \code{mia} package. For some reason,the names
are coming bad for some datasets. This is a workaround.
}
