% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_phyloseq}
\alias{filter_phyloseq}
\title{Fitler phyloseq object}
\usage{
filter_phyloseq(ps)
}
\arguments{
\item{ps}{A phyloseq object with otu_table and sample_data}
}
\value{
The filtered phyloseq object
}
\description{
\code{filter_phyloseq} filters both taxa and samples.
}
