% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_direction_cols}
\alias{get_direction_cols}
\title{Get direction columns}
\usage{
get_direction_cols(x, conditions_col, conditions)
}
\arguments{
\item{x}{Output of \code{benchdamic::runDA}.}

\item{conditions_col}{Character string indicating the name of the column
containing the conditions of the samples in sample_metadata/colData.}

\item{conditions}{A named character vector. The names must be
"condB" and "condA". condB indicates the reference/numerator/control
and condA indicates the target/denominator/treatment. For example:
\code{c(condB = 'control', condA = 'treatment')}}
}
\value{
A named vector ready to be used with the
\code{\link[benchdamic]{createEnrichment}} and
\code{\link[benchdamic]{createPositives}} functions.
}
\description{
\code{get_direction_cols} gets the names of the columns with the
effect sizes for each DA method. The output is suitable for
enrichment analyses with benchdamic.
}
