% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_meth_class}
\alias{get_meth_class}
\title{Methods classification}
\usage{
get_meth_class()
}
\value{
A tibble.
}
\description{
\code{method_classification} returns a tibble of the methods and their
classification.
}
