% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMethodClass}
\alias{getMethodClass}
\title{Get Method Classification}
\usage{
getMethodClass()
}
\value{
A tibble.
}
\description{
\code{getMethodClass} returns a tibble of the methods and their
classification. Helper function for plotting.
}
