% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefser2.R
\name{lefser2}
\alias{lefser2}
\title{R implementation of the LEfSe method 2}
\usage{
lefser2(
  expr,
  kruskal.threshold = 0.05,
  wilcox.threshold = 0.05,
  lda.threshold = 2,
  groupCol = "GROUP",
  blockCol = NULL,
  assay = 1L,
  trim.names = FALSE
)
}
\arguments{
\item{expr}{A \code{\linkS4class{SummarizedExperiment}} with expression data.}

\item{kruskal.threshold}{numeric(1) The p-value for the Kruskal-Wallis Rank
Sum Test (default 0.05).}

\item{wilcox.threshold}{numeric(1) The p-value for the Wilcoxon Rank-Sum Test
when 'blockCol' is present (default 0.05).}

\item{lda.threshold}{numeric(1) The effect size threshold (default 2.0).}

\item{groupCol}{character(1) Column name in \code{colData(expr)} indicating
groups, usually a factor with two levels (e.g., \code{c("cases", "controls")};
default "GROUP").}

\item{blockCol}{character(1) Optional column name in \code{colData(expr)}
indicating the blocks, usually a factor with two levels (e.g.,
\code{c("adult", "senior")}; default NULL).}

\item{assay}{The i-th assay matrix in the \code{SummarizedExperiment} ('expr';
default 1).}

\item{trim.names}{If \code{TRUE} extracts the most specific taxonomic rank of organism.}
}
\value{
The function returns a dataframe with two columns, which are
names of microorganisms and their LDA scores.
}
\description{
\code{lefser2} is an adaptation from the
\code{\link[lefser]{lefser}} function. This adaptation allows the inclusion
of the raw p-values of the KW test on the output.
}
