% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log2_fold_change}
\alias{log2_fold_change}
\title{Calculate log2 fold change}
\usage{
log2_fold_change(mat, condition_vector, ref = NULL)
}
\arguments{
\item{mat}{A matrix. Features in rows and samples in columns.}

\item{condition_vector}{A vector or factor with the names of the conditions.
Only two conditions are allowed and must be in the same order as the names
in the samples.}

\item{ref}{A reference for calculating the fold change. Default is NULL and
the first level of the factor will be used as reference.}
}
\value{
A named vector of log2 fold changes per feature.
}
\description{
\code{log2_fold_change} calculates the log2 fold change of the features
of a matrix per sample.
}
