% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_fold_change}
\alias{log_fold_change}
\title{Calculate log Fold Change}
\usage{
log_fold_change(num, denom)
}
\arguments{
\item{num}{vector with numerator}

\item{denom}{vector with denominator}
}
\value{
data transformed to log2 fold change
}
\description{
\code{log_fold_change} calculates log fold change
}
