% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{method_classification}
\alias{method_classification}
\title{Methods classification}
\usage{
method_classification()
}
\value{
A tibble.
}
\description{
\code{method_classification} returns a tibble of the methods and their
classification.
}
