% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{norm_CLR}
\alias{norm_CLR}
\title{CLR normalization}
\usage{
norm_CLR(mat, pseudocount = 0)
}
\arguments{
\item{pseudocount}{A pseudocount to add. Default = 1.}

\item{x}{A count matrix.}
}
\value{
A matrix with CLR normalization
}
\description{
\code{apply_clr} applies a centered-log ratio transformation to a matrix.
Features (e.g. taxa, OTUs) must be in the rows and samples in the columns.
}
