% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{norm_TSS}
\alias{norm_TSS}
\title{TSS normalization of matrix}
\usage{
norm_TSS(mat)
}
\arguments{
\item{mat}{A numeric matrix.}
}
\value{
A TSS-normalized matrix
}
\description{
\code{norm_TSS} Applies TSS normalization to a matrix of count data.
}
