% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_methods.R
\docType{data}
\name{normalization_methods}
\alias{normalization_methods}
\title{Normalization methods}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 2 columns.
}
\usage{
normalization_methods
}
\description{
Normalization methods
}
\keyword{datasets}
