% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_methods.R
\docType{data}
\name{normalization_methods_list}
\alias{normalization_methods_list}
\title{set normalization}
\format{
An object of class \code{list} of length 5.
}
\usage{
normalization_methods_list
}
\description{
set normalization
}
\keyword{datasets}
