% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Plot enrichment object}
\usage{
plot_enrichment(enrichment, enrichment_col, levels_to_plot, conditions)
}
\arguments{
\item{enrichment}{Enrichment object from benchdamic.}

\item{enrichment_col}{Column with enrichmnet annotations.}

\item{levels_to_plot}{Labels used for plotting. Default all.}

\item{conditions}{Named vector. condB first (i.e., reference, control).}
}
\value{
A ggplot object
}
\description{
\code{plot_enrichment} make plot from enrichment object.
}
