% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_enrichment_2}
\alias{plot_enrichment_2}
\title{Plot enrichment}
\usage{
plot_enrichment_2(data, dir)
}
\arguments{
\item{data}{A data.frame}

\item{dir}{Direction}
}
\value{
A ggplot object.
}
\description{
\code{plot_entichment_2}
}
