% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_positives}
\alias{plot_positives}
\title{Plot positives}
\usage{
plot_positives(x)
}
\arguments{
\item{x}{A dataframe}
}
\value{
List of plots
}
\description{
\code{plot_positives} is a version of plotPositives
}
