% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quiet}
\alias{quiet}
\title{Quiet}
\usage{
quiet(x)
}
\arguments{
\item{x}{Expression, command, etc.}
}
\value{
Nothing
}
\description{
\code{quiet} hides messages. This was taken from stackOverflow (reference
must be added).
}
