% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{relative_abundance_to_counts}
\alias{relative_abundance_to_counts}
\title{Convert from relative abundance (100) to counts}
\usage{
relative_abundance_to_counts(x, total_reads, total_sum = 100)
}
\arguments{
\item{x}{A matrix. Taxa in rownames. Samples in colnames.}

\item{total_reads}{Numeric vector with total reads per sample. It must be in
the same order as in the colnames of the matrix.}

\item{total_sum}{Total scaling. Default 100 (percentage).}
}
\value{
A matrix of counts
}
\description{
\code{relative_abundace_to_counts} converts from relative abundance to
counts.
}
