% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_DA.R
\name{run_DA}
\alias{run_DA}
\title{Run differential abundance methods}
\usage{
run_DA(object, conditions_col, conditions, verbose = FALSE)
}
\arguments{
\item{object}{A phyloseq object.}

\item{conditions_col}{A character string indicating the name of the column in
sample_metadata that contains the conditions to be compared.}

\item{conditions}{A named character vector. The names must be
"condB" and "condA". condB indicates the reference/numerator/control
and condA indicates the target/denominator/treatment. For example:
\code{c(condB = 'control', condA = 'treatment')}}

\item{verbose}{This argument is passed to the
\code{\link[benchdamic]{runNormalizations}} and
\code{\link[benchdamic]{runDA}} functions.}
}
\value{
A list containing the results of the DA methods. The output is
compatible with the benchdamic framework.
}
\description{
\code{run_DA} runs a set of predefined differential abundance methods on a
phyloseq object. This function includes normalization, calculation of matrix
of zinbweights, and running of the DA methods.
}
