% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_norm_list.R
\name{set_norm_list}
\alias{set_norm_list}
\title{Set list of normalization methods}
\usage{
set_norm_list()
}
\value{
A list of normalization methods compatible with the benchdamic
framework.
}
\description{
\code{set_norm_list} is a wrapper of the
\code{\link[benchdamic]{setNormalizations}} function.
}
\details{
The output should be used as input of the
\code{\link[benchdamic]{runNormalizations}} function.
}
