% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_methods.R
\docType{data}
\name{set_normalizations}
\alias{set_normalizations}
\title{set normalization}
\format{
An object of class \code{list} of length 5.
}
\usage{
set_normalizations
}
\description{
set normalization
}
\keyword{datasets}
