% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{taxize_classification_to_taxonomy_table}
\alias{taxize_classification_to_taxonomy_table}
\title{Taxize classification to a taxonomy table}
\usage{
taxize_classification_to_taxonomy_table(tax, id_type = "name")
}
\arguments{
\item{tax}{Output of \code{taxize::classification}.}

\item{id_type}{Either "name" or "id".}
}
\value{
A data frame
}
\description{
\code{taxize_classification_to_taxonomy_table} converts the output of
\code{taxize::classification} to a data frame.
}
\details{
This fuction was taken from the misctoolsr package on GitHub. Repo:
sdgamboa/misctoolsr
}
