% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchdamic_new_methods.R
\name{wilcox_test_clr}
\alias{wilcox_test_clr}
\title{Wilcox test with clr transformation}
\usage{
wilcox_test_clr(object, grp, ref = NULL, ...)
}
\arguments{
\item{object}{A phyloseq or (Tree)SummairizedExperiment object}

\item{grp}{A character string with the column name.}

\item{ref}{A character string with the condition used as reference.}

\item{...}{Other paramteres}
}
\value{
A list compatibble with the benchdamic workflow.
}
\description{
\code{wilcox_test_clr} is an adaptation of the wilcox.test function
(stats package) for use with the benchdamic workflow. A Centered-log ratio
transformation is applied to the matrix before performing the test.
}
