% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchdamic_new_methods.R
\name{zinq}
\alias{zinq}
\title{ZINQ for benchdamic}
\usage{
zinq(object, grp, ref = NULL, pval_type = "Cauchy", ...)
}
\arguments{
\item{object}{A (Tree)SummarizedExperiment or a phyloseq object.}

\item{grp}{The name of the grouping column, which is located in the colData
(SummarizedExperiment) or the sample_data (phyloseq).}

\item{ref}{A character string indicating the group that should be used as
reference.}

\item{pval_type}{A character string indicating the type of pvalue to use.
Options: Cauchy or MinP. Default: Cauchy.}

\item{...}{Other parameters}
}
\value{
A list in the format used in the benhcdamic pipeline.
}
\description{
\code{zinq} adapts the zinq method for integration with the \code{benchdamic}
package.
}
\section{References}{

Ling, W., Zhao, N., Plantinga, A.M. et al. Powerful and robust non-parametric
association testing for microbiome data via a zero-inflated quantile approach
(ZINQ). Microbiome 9, 181 (2021). https://doi.org/10.1186/s40168-021-01129-3
}

