% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_methods.R
\name{ancombc}
\alias{ancombc}
\title{ANCOMBC}
\usage{
ancombc(object, formula, group, ...)
}
\arguments{
\item{object}{A phyloseq object}

\item{formula}{A string.}

\item{group}{A string.}

\item{...}{Other parameters.}
}
\value{
ANCOMBC result.
}
\description{
\code{ancombc} performs ancombc
}
