% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_methods.R
\name{DA_ancombc}
\alias{DA_ancombc}
\title{DA_ancombc}
\usage{
DA_ancombc(
  object,
  pseudo_count = FALSE,
  norm = "none",
  conditions,
  verbose = TRUE,
  group,
  formula,
  ...
)
}
\arguments{
\item{object}{A phyloseq object.}

\item{pseudo_count}{Logical. Whether inlucde a pseudocount of 1 or not.
Default is FALSE.}

\item{norm}{Character string indicating the normalization to use.
Options: 'none' and 'TSS'. Default is 'none'.}

\item{conditions}{A named character vector of length 2 indicating the names
of the conditions to be compared. The names must be 'condB' for the reference
and 'condA' for the target, in that order. For example:
\code{c(condB = 'control', condA = 'treatment')}.}

\item{verbose}{Logical. If TRUE messages at each step are printed on screen.
Default is FALSE.}

\item{group}{Name of column with conditions/group information.
Same argument as in \code{\link[ANCOMBC]{ancombc}}. Check the original
function for more information.}

\item{formula}{Name of column with conditions/group information.
Same argument as in \code{\link[ANCOMBC]{ancombc}}. Check the original
function for more information.}

\item{...}{Parameters passed to the \code{\link[ANCOMBC]{ancombc}} function.}
}
\value{
A list with results of the analysis ready to be integrated in the
benchadmic framework.
}
\description{
\code{DA_ancombc} is an adaptation of the \code{\link[ANCOMBC]{ancombc}}
function for integration in the benchdamic framework.
}
\seealso{
\code{\link[ANCOMBC]{ancombc}}
}
