% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_methods_new.R
\name{DA_lefse}
\alias{DA_lefse}
\title{Lefser method}
\usage{
DA_lefse(object, grp, ref = NULL, ...)
}
\arguments{
\item{object}{A phyloseq or (Tree)SummarizedExperiment object.}

\item{grp}{A character string indicating the name of the column in colData
where the conditions are stored.}

\item{ref}{A character string indicating which condition should be used as
reference.}

\item{...}{Other parameters passed to benchdamic.}
}
\value{
An object for benchdamic.
}
\description{
\code{DA_lefser} is a modified version of the lefser package, which includes
the pvalues of the Kruskal test.
}
