% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_methods_new.R
\name{DA_wilcox}
\alias{DA_wilcox}
\title{Wilcox test for differential abundance}
\usage{
DA_wilcox(
  object,
  norm = "none",
  design,
  denom = NULL,
  pseudocount = NULL,
  verbose = FALSE
)
}
\arguments{
\item{object}{A phyloseq object or (Tree)SummarizedExperiment}

\item{norm}{String character. Choose between three normalization methods:
'none', 'CLR', or 'TSS'.}

\item{design}{String character.
The name of the column in sample data with the conditions.}

\item{denom}{The sample that will be used as denominator to calcluate
log2FoldChange. Usually, the 'treated' sample is used here.}

\item{verbose}{Print messages or not, Default is FALSE.}
}
\value{
A list with outputs compatible with the benchdamic framework/package.
}
\description{
\code{DA_wilcox} performs Wilcoxon test on a phyloseq object.
}
