% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_phyloseq}
\alias{filter_phyloseq}
\title{Fitler phyloseq object}
\usage{
filter_phyloseq(ps, min_per_samples = 0.2)
}
\arguments{
\item{ps}{A phyloseq object with otu_table and sample_data}

\item{min_per_samples}{minimal percentage of samples}
}
\value{
The filtered phyloseq object
}
\description{
\code{filter_phyloseq} filters both taxa and samples.
}
