% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_methods.R
\name{log2_fold_change}
\alias{log2_fold_change}
\title{Calculate log2 fold change}
\usage{
log2_fold_change(
  mat,
  condition_vector,
  condB = NULL,
  log = FALSE,
  pseudocount = 0
)
}
\arguments{
\item{mat}{A matrix. Features in rows and samples in columns.}

\item{condition_vector}{A character vector or factor with the names of the
conditions. The conditions must correspond to the samples, i.e. the exact
same order. Two and only two conditions (levels) are needed.}

\item{condB}{Condition used as reference. E.g. control condition.}

\item{log}{If log is TRUE, it's assumed that the matrix is already log
transformed.}

\item{pseudocount}{Numeric value indicating pseudocount to be added.}
}
\value{
A named vector of log2 fold changes per feature.
}
\description{
\code{log2_fold_change} calculates the log2 fold change of a matrix with
features in the rows and samples in the columns.
}
