% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeBenchmarkDataAnalyses.R
\name{MicrobiomeBenchmarkDataAnalyses}
\alias{MicrobiomeBenchmarkDataAnalyses}
\title{MicrobiomeBenchmarkDataAnalyses}
\description{
\code{MicrobiomeBenchmarkDataAnalyses} contains the code used
to analyze the microbiome datasets in the \code{MicrobiomeBenchmarkData}
package.
}
