% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeBenchmarkDataAnalyses.R
\docType{package}
\name{MicrobiomeBenchmarkDataAnalyses}
\alias{MicrobiomeBenchmarkDataAnalyses}
\title{MicrobiomeBenchmarkDataAnalyses}
\description{
\code{MicrobiomeBenchmarkDataAnalyses} is a package for the DA analyses
performed on the MicrobiomeBenchmarkData package.
}
