% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{norm_clr}
\alias{norm_clr}
\title{CLR normalization}
\usage{
norm_clr(mat, pseudocount = 0, log = TRUE)
}
\arguments{
\item{mat}{A count matrix.}

\item{pseudocount}{Pseudocount added to the matrix.
Default value is 0 (no pseudocount)..}

\item{log}{If TRUE, CLR will be logged. Default = TRUE. In most
cases this should be set to TRUE}
}
\value{
A CLR-transformed matrix.
}
\description{
\code{norm_clr} applies a centered-log ratio (CLR) transformation to a
matrix column-wise. Features (e.g. taxa, OTUs) must be in the rows and
samples in the columns.
}
