% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_methods.R
\name{set_DA_methods_list}
\alias{set_DA_methods_list}
\title{Set list of DA methods}
\usage{
set_DA_methods_list(grp, contrast)
}
\arguments{
\item{grp}{Character string. Name of the column}

\item{contrast}{Character vector. Reference goes first.}
}
\value{
A list of methods for benchdamic
}
\description{
\code{set_DA_methods_list} creates a list of methods for the benchdamic
workflow.
}
