% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_DA_methods_list.R
\name{set_DA_methods_list}
\alias{set_DA_methods_list}
\title{Set list of DA methods}
\usage{
set_DA_methods_list(conditions_col, conditions)
}
\arguments{
\item{conditions_col}{Character string indicating the name of the column
containing the conditions of the samples in sample_metadata/colData.}
}
\value{
A list of DA methods for benchdamic.
}
\description{
\code{set_DA_methods_list} creates a predefined list of methods for the
benchdamic framework. The output of this function should input for the
\code{\link[benchdamic]{runDA}} function.
}
